fasta_file_creator/requirements.txt
pandas==1.3.4
biopython==1.79
